requirements.txt
python-dotenv
expertai-nlapi
fastapi
uvicorn[standard]
yfinancec
httpx
psycopg2
configparser
pandas
pygooglenews
newspaper3k